flash messages

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController {
   
@@ -22,7 +24,7 @@ class ListingController {
    */
   public function index(){
 
-    $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+    $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
 loadView('listings/index', [
   'listings' => $listings
@@ -74,7 +76,7 @@ loadView('listings/index', [
 
     $newListingData = array_map('sanitize', $newListingData);
 
-    $newListingData['user_id'] = 1;
+    $newListingData['user_id'] = Session::get('user')['id'];
     
     $requiredFields = ['title','description', 'salary', 'city', 'state',  'email'  ];
     
@@ -118,6 +120,9 @@ loadView('listings/index', [
 
   $this->db->query($query, $newListingData);
 
+  Session::setFlashMessage('success_message', 'Listing created successfully');
+  return redirect('/listings');
+
   redirect("/listings");
 }
 }
@@ -137,15 +142,23 @@ public function destroy($params){
   ];
 $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+// Check if listing exists
 if (!$listing) {
   ErrorController::notFound('Listing not found');
   return;
 }
+// Authorization
+if(!Authorization::isOwner($listing->user_id)){
+
+  Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+  return redirect('/listings/' . $listing->id);
+}
 
 $this->db->query('DELETE FROM listings WHERE id = :id', $params);
+
 //Set flash message
-$_SESSION['success_message'] = 'Listing deleted successfully';
-redirect('/listings');
+Session::setFlashMessage('success_message', 'Listing deleted successfully');
+  return redirect('/listings');
 }
   /**
    * Show the listing edit form
